==> 1. SEC_R Anonymization&ReverseCoding.rmd <==
---
title: "SEC_R Anonymization&ReverseCoding"
output: html_document
---
#readr for reading csv files as tibble
#stargazer for tables
#car for reverse coding
#psych for internal consistency
#VIM for missing data graphs
#mice for missing data imputation and exploration
#summarytools for missing data, histograms, mean, and SD
#}

#Clear memory
rm(list=ls())
#}

#load libraries and data {
library(stargazer)
library (car)
library(summarytools)
library(VIM)
library (psych)
library(tidyverse)
library(readr)
library(rowr)
library(lavaan)

#Load local identified data, assign to tibble
tib_identified <- read_csv(file = "C:/Users/dalla/Google Drive/CURRENT/SEC_R_SGISaskGaming/SEC_R_crowdsourced/SEC_R_Study_CrowdSourced_February 4, 2019_15.08_numeric.csv")
#}

names(tib_identified)[names(tib_identified)=="Duration (in seconds)"] <- "Duration"
names(tib_identified)[names(tib_identified)=="Recorded Date"] <- "RecordedDate"


tib <- dplyr::select (tib_identified, -c(StartDate, EndDate, Status, IPAddress, Progress, Duration, Finished, RecordedDate, ResponseId, RecipientLastName, RecipientFirstName, RecipientEmail, ExternalReference, LocationLatitude, LocationLongitude, DistributionChannel, UserLanguage, ResultsEmail))

names(tib)[names(tib)=="pincome"] <- "Pincome"

write.csv(tib, file="C:/Users/dalla/Google Drive/R Coding/Comparisons_Risk/SEC_R_Crowd_2_24_2019_anonymized.csv")

library(RCurl)

x <- getURL("https://raw.githubusercontent.com/DallasNovakowski/Comparisons_Risk/master/SEC_R_Crowd_2_24_2019_anonymized.csv")

tib_cleaning <- read.csv(text=x)

#reverse coding for impulsivity and sensation-seeking scales
tib_cleaning$BIS_6r <- as.numeric(dplyr::recode(tib_cleaning$BIS_6, "1" = "4" , "3" = "2" , "4" = "1" , "2" = "3"))
tib_cleaning$BIS_7r <- as.numeric(dplyr::recode(tib_cleaning$BIS_7, "1" = "4" , "3" = "2" , "4" = "1" , "2" = "3"))
tib_cleaning$BIS_8r <- as.numeric(dplyr::recode(tib_cleaning$BIS_8, "1" = "4" , "3" = "2" , "4" = "1" , "2" = "3"))
tib_cleaning$BIS_9r <- as.numeric(dplyr::recode(tib_cleaning$BIS_9, "1" = "4" , "3" = "2" , "4" = "1" , "2" = "3"))
tib_cleaning$BIS_10r <- as.numeric(dplyr::recode(tib_cleaning$BIS_10, "1" = "4" , "3" = "2" , "4" = "1" , "2" = "3"))
tib_cleaning$BIS_13r <- as.numeric(dplyr::recode(tib_cleaning$BIS_13, "1" = "4" , "3" = "2" , "4" = "1" , "2" = "3"))


tib_cleaning$BSCS_2r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_2, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))
tib_cleaning$BSCS_3r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_3, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))
tib_cleaning$BSCS_4r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_4, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))
tib_cleaning$BSCS_5r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_5, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))
tib_cleaning$BSCS_7r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_7, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))
tib_cleaning$BSCS_9r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_9, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))
tib_cleaning$BSCS_10r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_10, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))
tib_cleaning$BSCS_12r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_12, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))
tib_cleaning$BSCS_13r <- as.numeric(dplyr::recode(tib_cleaning$BSCS_13, "1" = "5" , "2" = "4" , "3" = "3" , "4" = "2" , "5" = "1"))

#reverse-code SEC_7: "I feel that I am a person of worth, at least on an equal plane with others. "
tib_cleaning$SEC_7r <- as.numeric(dplyr::recode(tib_cleaning$SEC_7, "1" = "7" , "2" = "6", "3" = "5", "4" = "4", "5" = "3", "6" = "2", "7" = "1"))

#recoding FTND measures because they're skewed by +1
tib_cleaning$FTND_1f <- tib_cleaning$FTND_1 -1 
tib_cleaning$FTND_2f <- tib_cleaning$FTND_2 -1 
tib_cleaning$FTND_3f <- tib_cleaning$FTND_3 -1 
tib_cleaning$FTND_4f <- tib_cleaning$FTND_4 -1 
tib_cleaning$FTND_5f <- tib_cleaning$FTND_5 -1 
tib_cleaning$FTND_6f <- tib_cleaning$FTND_6 -1 

#recoding NODS measures so that 1(yes)=1 and 2(no)=0 
tib_cleaning$NODS_1f <- as.numeric(dplyr::recode(tib_cleaning$NODS_1, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_2f <- as.numeric(dplyr::recode(tib_cleaning$NODS_2, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_3f <- as.numeric(dplyr::recode(tib_cleaning$NODS_3, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_4f <- as.numeric(dplyr::recode(tib_cleaning$NODS_4, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_5f <- as.numeric(dplyr::recode(tib_cleaning$NODS_5, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_6f <- as.numeric(dplyr::recode(tib_cleaning$NODS_6, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_7f <- as.numeric(dplyr::recode(tib_cleaning$NODS_7, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_8f <- as.numeric(dplyr::recode(tib_cleaning$NODS_8, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_9f <- as.numeric(dplyr::recode(tib_cleaning$NODS_9, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_10f <- as.numeric(dplyr::recode(tib_cleaning$NODS_10, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_11f <- as.numeric(dplyr::recode(tib_cleaning$NODS_11, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_12f <- as.numeric(dplyr::recode(tib_cleaning$NODS_12, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_13f <- as.numeric(dplyr::recode(tib_cleaning$NODS_13, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_14f <- as.numeric(dplyr::recode(tib_cleaning$NODS_14, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_15f <- as.numeric(dplyr::recode(tib_cleaning$NODS_15, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_16f <- as.numeric(dplyr::recode(tib_cleaning$NODS_16, "1" = "1" , "2" = "0"))
tib_cleaning$NODS_17f <- as.numeric(dplyr::recode(tib_cleaning$NODS_17, "1" = "1" , "2" = "0"))

#recoding DAST measures so that 1(yes)=1 and 2(no)=0 (and reverse-code DAST_3f)
tib_cleaning$DAST_1f <- as.numeric(dplyr::recode(tib_cleaning$DAST_1, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_2f <- as.numeric(dplyr::recode(tib_cleaning$DAST_2, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_3f <- as.numeric(dplyr::recode(tib_cleaning$DAST_3, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_3fr <- as.numeric(dplyr::recode(tib_cleaning$DAST_3f, "1" = "0" , "0" = "1"))
tib_cleaning$DAST_4f <- as.numeric(dplyr::recode(tib_cleaning$DAST_4, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_5f <- as.numeric(dplyr::recode(tib_cleaning$DAST_5, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_6f <- as.numeric(dplyr::recode(tib_cleaning$DAST_6, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_7f <- as.numeric(dplyr::recode(tib_cleaning$DAST_7, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_8f <- as.numeric(dplyr::recode(tib_cleaning$DAST_8, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_9f <- as.numeric(dplyr::recode(tib_cleaning$DAST_9, "1" = "1" , "2" = "0"))
tib_cleaning$DAST_10f <- as.numeric(dplyr::recode(tib_cleaning$DAST_10, "1" = "1" , "2" = "0"))




#create lists for variable computation
VarBase <- list()
VarBase$Pnum <- c('Pnum', 'AttnChckPass')
VarBase$demo <-  c('age', 'sex', 'Pincome', 'Hincome', 'Smoke_YN')
VarBase$SSexp <-  c('SS_1', 'SS_5')
VarBase$SSdis <-  c('SS_4', 'SS_8')
VarBase$SSas <-  c('SS_3', 'SS_7')
VarBase$SSbor <-  c('SS_2', 'SS_6')
VarBase$SStot <- c('SS_1', 'SS_2', 'SS_3', 'SS_4','SS_5', 'SS_6', 'SS_7', 'SS_8')
VarBase$BISa <-  c('BIS_11', 'BIS_12', 'BIS_13r', 'BIS_14', 'BIS_15')
VarBase$BISm <-  c('BIS_1', 'BIS_2', 'BIS_3', 'BIS_4', 'BIS_5')
VarBase$BISn <-  c('BIS_6r', 'BIS_7r', 'BIS_8r', 'BIS_9r', 'BIS_10r')
VarBase$Dhea <-  c('DsrA_5', 'DsrA_15', 'DsrA_17', 'DsrA_20', 'DsrA_23', 'DsrA_26')
VarBase$Dinv <-  c('DsrA_4', 'DsrA_12', 'DsrA_18')
VarBase$Dgam <-  c('DsrA_3', 'DsrA_8', 'DsrA_14')
VarBase$Deth <-  c('DsrA_6', 'DsrA_9', 'DsrA_10', 'DsrA_16', 'DsrA_29', 'DsrA_30')
VarBase$Drec <-  c('DsrA_2', 'DsrA_11', 'DsrA_13', 'DsrA_19', 'DsrA_24', 'DsrA_25')
VarBase$Dsoc <-  c('DsrA_1', 'DsrA_7', 'DsrA_21', 'DsrA_22', 'DsrA_27', 'DsrA_28')
VarBase$SOEP <-  c('SOEPg', 'SOEPdri', 'SOEPocc', 'SOEPfin', 'SOEPrec', 'SOEPhea', 'SOEPsoc')
VarBase$Dm <-  c('Dm_1', 'Dm_2', 'Dm_3', 'Dm_4', 'Dm_5', 'Dm_6', 'Dm_7', 'Dm_8', 'Dm_9', 'Dm_10', 'Dm_11', 'Dm_12', 'Dm_13', 'Dm_14', 'Dm_15', 'Dm_16', 'Dm_17', 'Dm_18', 'Dm_19')
VarBase$GBS <-  c('GBS_lott', 'GBS_raf', 'GBS_scrat', 'GBS_bing', 'GBS_vlt', 'GBS_table', 'GBS_sport', 'GBS_horse', 'GBS_skill', 'GBS_stock')
VarBase$GBS_all <-  c('GBS_lott', 'GBS_raf', 'GBS_scrat', 'GBS_bing', 'GBS_vlt', 'GBS_table', 'GBS_sport', 'GBS_horse', 'GBS_skill', 'GBS_stock', 'GBS_oth1', 'GBS_oth2')
VarBase$DASTb <-  c('DAST_1', 'DAST_2', 'DAST_3', 'DAST_4', 'DAST_5', 'DAST_6', 'DAST_7', 'DAST_8', 'DAST_9', 'DAST_10')
VarBase$DAST <-  c('DAST_1f', 'DAST_2f', 'DAST_3fr', 'DAST_4f', 'DAST_5f', 'DAST_6f', 'DAST_7f', 'DAST_8f', 'DAST_9f', 'DAST_10f')
VarBase$FTND <-  c('FTND_1f', 'FTND_2f', 'FTND_3f', 'FTND_4f', 'FTND_5f', 'FTND_6f')
VarBase$AUDIT <-  c('AUDIT_1', 'AUDIT_2', 'AUDIT_3', 'AUDIT_4', 'AUDIT_5', 'AUDIT_6', 'AUDIT_7', 'AUDIT_8', 'AUDIT_9', 'AUDIT_10')
VarBase$NODS <-  c('NODS_1f', 'NODS_2f', 'NODS_3f', 'NODS_4f', 'NODS_5f', 'NODS_6f', 'NODS_7f', 'NODS_8f', 'NODS_9f', 'NODS_10f', 'NODS_11f', 'NODS_12f', 'NODS_13f', 'NODS_14f', 'NODS_16f', 'NODS_17f')
#pipe = only presented if condition met; 7: if unsuccessful cessation attempted, 12: if lied to social network, 15: only if in school
VarBase$NODSpipe <- c('NODS_7f', 'NODS_12f','NODS_15f')
VarBase$SECtot <-  c('SEC_1', 'SEC_2', 'SEC_3', 'SEC_4', 'SEC_5', 'SEC_6', 'SEC_7r', 'SEC_8', 'SEC_9', 'SEC_10', 'SEC_11', 'SEC_12')
VarBase$SECenvy <-  c('SEC_1', 'SEC_2', 'SEC_3', 'SEC_4')
VarBase$SECcse <-  c('SEC_5', 'SEC_6','SEC_7r', 'SEC_8')
VarBase$SECjs <- c('SEC_9','SEC_10', 'SEC_11', 'SEC_12')

VarBase$ALL <- c(VarBase$SSexp, VarBase$SSdis, VarBase$SSas, VarBase$SSbor, VarBase$BISa, VarBase$BISm, VarBase$BISn, VarBase$Dhea, VarBase$Dinv, VarBase$Dgam, VarBase$Deth, VarBase$Drec, VarBase$Dsoc, VarBase$SOEP, VarBase$Dm, VarBase$GBS, VarBase$NODS, VarBase$DAST, VarBase$AUDIT, VarBase$SECtot)

# Recode NODS_7, NODS_12, NODS_15 so that NA = 0 (piped items that are skipped with "no" responses)
tib_cleaning[,c(VarBase$NODSpipe)][is.na(tib_cleaning[,c(VarBase$NODSpipe)])] <- 0

#create cleaning datasets, only including those participants that passed the attention check
tib_cleaning <- filter(tib_cleaning, tib_cleaning$AttnChckPass == 1)
tib_imputed <- filter(tib_cleaning, tib_cleaning$AttnChckPass == 1)

#only include participants that rated >= 5 on the opening attention check item (out of 7)
tib_cleaning <- filter(tib_cleaning, tib_cleaning$AttnChck_rate >= 5)
tib_imputed <- filter(tib_cleaning, tib_cleaning$AttnChck_rate >= 5)

#Give nonsmokers 0 scores on all FTND items
tib_cleaning[,c(VarBase$FTND)][tib_cleaning$Smoke_YN == 2 & is.na(tib_cleaning[,c(VarBase$FTND)])] <- 0
tib_imputed[,c(VarBase$FTND)][tib_imputed$Smoke_YN == 2 & is.na(tib_imputed[,c(VarBase$FTND)])] <- 0

# remove Pnums 91, 606, and 606 due to excessive missing data
tib_cleaning <- tib_cleaning[tib_cleaning$Pnum!=91 & tib_cleaning$Pnum!=966 & tib_cleaning$Pnum!=606, ]


#tib_cleaning <- dplyr::select (tib_identified, -c(StartDate, EndDate, Status, IPAddress, Progress, Duration, Finished, RecordedDate, ResponseId, RecipientLastName, RecipientFirstName, RecipientEmail, ExternalReference, LocationLatitude, LocationLongitude, DistributionChannel, UserLanguage, ResultsEmail))


write.csv(tib_cleaning, file="C:/Users/dalla/Google Drive/R Coding/Comparisons_Risk/SEC_R_Crowd_2_24_2019_Rcoded.csv")
