==> 2. SEC_R ImputationAndComputation.rmd <==
---
title: "2. SEC_R ImputationAndComputation"
output: html_document
---

#If you just want the resulting dataset without having to go through the lengthy imputation
load(file="C:/Users/dalla/Google Drive/R Coding/Comparisons_Risk/SEC_R_Crowd_2_22_2019_imputed.Rdata")

library(RCurl)

x <- getURL("https://raw.githubusercontent.com/DallasNovakowski/Comparisons_Risk/master/SEC_R_Crowd_2_24_2019_Rcoded.csv")

tib_cleaning <- read.csv(text=x)

tib_cleaned <- tib_cleaning

Let's see if age needs cleaning 
```{r Age before cleaning}
hist(tib_cleaning$age)
```


```{r Sex before cleaning}
hist(tib_cleaning$sex)
ind <- which(tib_cleaning$sex > 2)
tib_cleaning[ind,]
#8 participants either responded "other"" or did not disclose
other <- which(tib_cleaning$sex == 3)
nodisclose <- which(tib_cleaning$sex == 4)
```

# how many other? (4); 2 are genuine queer; the remaining 2 are nonsensical ("heterosexual")
NROW(other)
``` {r}
tib_cleaned[other,]
```

#let's drop the nonsensical responses from the cleaned data frame

tib_cleaned <- subset(tib_cleaned, !Pnum%in%c(116, 682)) 
#there are 4 no-disclose;  they (and the legitimate "other" sex responses) will be retained for now (although they will end up being excluded from any analyses that use the ~ age+sex residuals)
NROW(nodisclose)
tib_cleaning[nodisclose,]

``` {r}
hist(tib_cleaned$sex)
```



#imputing all items except demographics, and assigning these imputed datasets to unique objects
set.seed(111)
dat_na <- tib_cleaned[,c(VarBase$SSexp, VarBase$SSdis, VarBase$SSas, VarBase$SSbor, VarBase$BISa, VarBase$BISm, VarBase$BISn, VarBase$Dhea, VarBase$Dinv, VarBase$Dgam, VarBase$Deth, VarBase$Drec, VarBase$Dsoc, VarBase$SOEP, VarBase$Dm, VarBase$GBS, VarBase$NODS, VarBase$FTND, VarBase$DAST, VarBase$AUDIT, VarBase$SECtot)]

library (mice)

#WARNING: THIS IMPUTATION PROCEDURE TAKES A VERY LONG TIME
imputed <- mice(dat_na, m=20, maxit=10, meth='pmm', seed=500)

  dat_imp1 <- mice::complete(imputed, 1)
	dat_imp2 <- mice::complete(imputed, 2)
	dat_imp3 <- mice::complete(imputed, 3)
	dat_imp4 <- mice::complete(imputed, 4)
	dat_imp5 <- mice::complete(imputed, 5)
	dat_imp6 <- mice::complete(imputed, 6)
	dat_imp7 <- mice::complete(imputed, 7)
	dat_imp8 <- mice::complete(imputed, 8)
	dat_imp9 <- mice::complete(imputed, 9)
	dat_imp10 <- mice::complete(imputed, 10)
	dat_imp11 <- mice::complete(imputed, 11)
	dat_imp12 <- mice::complete(imputed, 12)
	dat_imp13 <- mice::complete(imputed, 13)
	dat_imp14 <- mice::complete(imputed, 14)
	dat_imp15 <- mice::complete(imputed, 15)
	dat_imp16 <- mice::complete(imputed, 16)
	dat_imp17 <- mice::complete(imputed, 17)
	dat_imp18 <- mice::complete(imputed, 18)
	dat_imp19 <- mice::complete(imputed, 19)
	dat_imp20 <- mice::complete(imputed, 20)

#investigate logged events from imputation
imputed$loggedEvents

#Binding participant numbers to imputed data sets
dat_imp1 <- cbind.fill(dat_imp1,tib_cleaned[,c("Pnum")])
dat_imp2 <- cbind.fill(dat_imp2,tib_cleaned[,c("Pnum")])
dat_imp3 <- cbind.fill(dat_imp3,tib_cleaned[,c("Pnum")])
dat_imp4 <- cbind.fill(dat_imp4,tib_cleaned[,c("Pnum")])
dat_imp5 <- cbind.fill(dat_imp5,tib_cleaned[,c("Pnum")])
dat_imp6 <- cbind.fill(dat_imp6,tib_cleaned[,c("Pnum")])
dat_imp7 <- cbind.fill(dat_imp7,tib_cleaned[,c("Pnum")])
dat_imp8 <- cbind.fill(dat_imp8,tib_cleaned[,c("Pnum")])
dat_imp9 <- cbind.fill(dat_imp9,tib_cleaned[,c("Pnum")])
dat_imp10 <- cbind.fill(dat_imp10,tib_cleaned[,c("Pnum")])
dat_imp11 <- cbind.fill(dat_imp11,tib_cleaned[,c("Pnum")])
dat_imp12 <- cbind.fill(dat_imp12,tib_cleaned[,c("Pnum")])
dat_imp13 <- cbind.fill(dat_imp13,tib_cleaned[,c("Pnum")])
dat_imp14 <- cbind.fill(dat_imp14,tib_cleaned[,c("Pnum")])
dat_imp15 <- cbind.fill(dat_imp15,tib_cleaned[,c("Pnum")])
dat_imp16 <- cbind.fill(dat_imp16,tib_cleaned[,c("Pnum")])
dat_imp17 <- cbind.fill(dat_imp17,tib_cleaned[,c("Pnum")])
dat_imp18 <- cbind.fill(dat_imp18,tib_cleaned[,c("Pnum")])
dat_imp19 <- cbind.fill(dat_imp19,tib_cleaned[,c("Pnum")])
dat_imp20 <- cbind.fill(dat_imp20,tib_cleaned[,c("Pnum")])

#renaming object (which pnum got named as in cbind function) to "pnum"
library(plyr)
dat_imp1 <- rename(dat_imp1, c("object"="Pnum"))
dat_imp2 <- rename(dat_imp2, c("object"="Pnum"))
dat_imp3 <- rename(dat_imp3, c("object"="Pnum"))
dat_imp4 <- rename(dat_imp4, c("object"="Pnum"))
dat_imp5 <- rename(dat_imp5, c("object"="Pnum"))
dat_imp6 <- rename(dat_imp6, c("object"="Pnum"))
dat_imp7 <- rename(dat_imp7, c("object"="Pnum"))
dat_imp8 <- rename(dat_imp8, c("object"="Pnum"))
dat_imp9 <- rename(dat_imp9, c("object"="Pnum"))
dat_imp10 <- rename(dat_imp10, c("object"="Pnum"))
dat_imp11 <- rename(dat_imp11, c("object"="Pnum"))
dat_imp12 <- rename(dat_imp12, c("object"="Pnum"))
dat_imp13 <- rename(dat_imp13, c("object"="Pnum"))
dat_imp14 <- rename(dat_imp14, c("object"="Pnum"))
dat_imp15 <- rename(dat_imp15, c("object"="Pnum"))
dat_imp16 <- rename(dat_imp16, c("object"="Pnum"))
dat_imp17 <- rename(dat_imp17, c("object"="Pnum"))
dat_imp18 <- rename(dat_imp18, c("object"="Pnum"))
dat_imp19 <- rename(dat_imp19, c("object"="Pnum"))
dat_imp20 <- rename(dat_imp20, c("object"="Pnum"))



#making summary imputed data set (taking mean from the 20 separate imputed data sets), then adding demographics
#NOTE: this isn't best practice, I should be running analyses for all 20 of these imputed datasets, then average the values of the parameter estimates to get a single point estimate. I DIDN'T do this because...
#there's also the question of whether I should have conducted imputations for summed scales/subscales or for the item-level indicators (I conducted the latter)
library(data.table)

tib_imputed <- rbindlist(list(dat_imp1, dat_imp2, dat_imp3, dat_imp4, dat_imp5, dat_imp6, dat_imp7, dat_imp8, dat_imp9, dat_imp10, dat_imp11, dat_imp12, dat_imp13, dat_imp14, dat_imp15, dat_imp16, dat_imp17, dat_imp18, dat_imp19, dat_imp20))[,lapply(.SD,mean), Pnum]

#drop NODS_15f because everyone answered "no"

tib_filled <- cbind.fill(tib_imputed,tib_cleaned[,c(VarBase$demo)])


# calculate Cronbach's alphas
c_alpha <- list()
c_alpha$SOEP <- psych:: alpha(tib_filled[,c(VarBase$SOEP)])
c_alpha$SSexp <- psych:: alpha(tib_filled[,c(VarBase$SSexp)])
c_alpha$SSdis <- psych:: alpha(tib_filled[,c(VarBase$SSdis)])
c_alpha$SSas <- psych:: alpha(tib_filled[,c(VarBase$SSas)])
c_alpha$SSbor <- psych:: alpha(tib_filled[,c(VarBase$SSbor)])
c_alpha$SStot <- psych:: alpha(tib_filled[,c(VarBase$SStot)])
c_alpha$BIStot <- psych:: alpha(tib_filled[,c(VarBase$BISa, VarBase$BISm, VarBase$BISn)])
c_alpha$BISa <- psych:: alpha(tib_filled[,c(VarBase$BISa)])
c_alpha$BISm <- psych:: alpha(tib_filled[,c(VarBase$BISm)])
c_alpha$BISn <- psych:: alpha(tib_filled[,c(VarBase$BISn)])
c_alpha$Dhea <- psych:: alpha(tib_filled[,c(VarBase$Dhea)])
c_alpha$Dinv <- psych:: alpha(tib_filled[,c(VarBase$Dinv)])
c_alpha$Dgam <- psych:: alpha(tib_filled[,c(VarBase$Dgam)])
c_alpha$Deth <- psych:: alpha(tib_filled[,c(VarBase$Deth)])
c_alpha$Drec <- psych:: alpha(tib_filled[,c(VarBase$Drec)])
c_alpha$Dsoc <- psych:: alpha(tib_filled[,c(VarBase$Dsoc)])
c_alpha$SOEP <- psych:: alpha(tib_filled[,c(VarBase$SOEP)])
c_alpha$Dm <- psych:: alpha(tib_filled[,c(VarBase$Dm)])
c_alpha$GBS <- psych:: alpha(tib_filled[,c(VarBase$GBS)])
c_alpha$GBS_all <- psych:: alpha(tib_filled[,c(VarBase$GBS_all)])
c_alpha$DASTb <- psych:: alpha(tib_filled[,c(VarBase$DASTb)])
c_alpha$DAST <- psych:: alpha(tib_filled[,c(VarBase$DAST)])
c_alpha$FTND <- psych:: alpha(tib_filled[,c(VarBase$FTND)])
c_alpha$AUDIT <- psych:: alpha(tib_filled[,c(VarBase$AUDIT)])
c_alpha$NODS <- psych:: alpha(tib_filled[,c(VarBase$NODS)])
c_alpha$NODSpipe <- psych:: alpha(tib_filled[,c(VarBase$NODSpipe)])
c_alpha$SECtot <- psych:: alpha(tib_filled[,c(VarBase$SECtot)])
c_alpha$SECenvy <- psych:: alpha(tib_filled[,c(VarBase$SECenvy)])
c_alpha$SECcse <- psych:: alpha(tib_filled[,c(VarBase$SECcse)])
c_alpha$SECjs <- psych:: alpha(tib_filled[,c(VarBase$SECjs)])


#create a list of raw alpha values from the c_alpha list, which took like an hour because I'm the smartest stupid person on the planet
alphas <- lapply(c_alpha, FUN = function(x){
    x$total$raw_alpha
})

# VARIABLE COMPUTATION
SSexp <- rowSums(tib_filled[,c(VarBase$SSexp)])
SSdis <- rowSums(tib_filled[,c(VarBase$SSdis)])
SSas <- rowSums(tib_filled[,c(VarBase$SSas)])
SSbor <- rowSums(tib_filled[,c(VarBase$SSbor)])
BISa <- rowSums(tib_filled[,c(VarBase$BISa)])
BISm <- rowSums(tib_filled[,c(VarBase$BISm)])
BISn <- rowSums(tib_filled[,c(VarBase$BISn)])
Dhea <- rowSums(tib_filled[,c(VarBase$Dhea)])
Dinv <- rowSums(tib_filled[,c(VarBase$Dinv)])
Dgam <- rowSums(tib_filled[,c(VarBase$Dgam)])
Deth <- rowSums(tib_filled[,c(VarBase$Deth)])
Drec <- rowSums(tib_filled[,c(VarBase$Drec)])
Dsoc <- rowSums(tib_filled[,c(VarBase$Dsoc)])
Dm <- rowSums(tib_filled[,c(VarBase$Dm)])
GBS <- rowSums(tib_filled[,c(VarBase$GBS)])
DAST <- rowSums(tib_filled[,c(VarBase$DAST)])
FTND <- rowSums(tib_filled[,c(VarBase$FTND)])
AUDIT <- rowSums(tib_filled[,c(VarBase$AUDIT)])
NODS <- rowSums(tib_filled[,c(VarBase$NODS)])
SECtot <- rowSums(tib_filled[,c(VarBase$SECtot)])
SECenvy <- rowSums(tib_filled[,c(VarBase$SECenvy)])
SECcse <- rowSums(tib_filled[,c(VarBase$SECcse)])
SECjs <- rowSums(tib_filled[,c(VarBase$SECjs)])

#all variables that are the result of a summation process
VarBase$Summed <- c('SSexp', 'SSdis', 'SSas', 'SSbor', 'BISa', 'BISm', 'BISn', 'Dhea', 'Dinv', 'Dgam', 'Deth', 'Drec', 'Dsoc', 'Dm', 'GBS', 'DAST', 'FTND', 'AUDIT', 'NODS', 'SECenvy', 'SECcse', 'SECjs')


#adding computed variables to new dataset
tib_summed <- cbind(tib_filled, SSexp, SSdis, SSas, SSbor, BISa, BISm, BISn, Dhea, Dinv, Dgam, Deth, Drec, Dsoc, Dm, GBS, DAST, FTND, AUDIT, NODS, SECenvy, SECcse, SECjs)

hist(tib_summed$SSexp)
hist(tib_summed$SSdis)
hist(tib_summed$SSas)
hist(tib_summed$SSbor)
hist(tib_summed$BISa)
hist(tib_summed$BISm)
hist(tib_summed$BISn)
hist(tib_summed$Dhea)
hist(tib_summed$Dinv)
#has large skew
hist(tib_summed$Dgam)

hist(tib_summed$Deth)
hist(tib_summed$Drec)
hist(tib_summed$Dsoc)
hist(tib_summed$Dm)

#also has large skew
hist(tib_summed$GBS)

# also also has large skew
hist(tib_summed$AUDIT)

hist(tib_summed$SECenvy)
hist(tib_summed$SECcse)
hist(tib_summed$SECjs)

#Extremely Skewed
hist(tib_summed$FTND)
hist(tib_summed$DAST)
hist(tib_summed$NODS)


# dichotomization of extremely skewed pathological behaviours
tib_summed$FTND_dich <- bin(tib_summed$FTND, nbins = round(nrow(tib_summed)/50), labels = NULL, method = c("length", "content", "clusters"), na.omit = TRUE)

tib_summed$FTND_dich <- NA

tib_summed$FTND_dich <- as.numeric(dplyr::recode(tib_summed$FTND, "0" = "0" , "1" = "1", "2" = "1", "3" = "1", "4" = "1", "5" = "1", "6" = "1", "7" = "1", "8" = "1", "9" = "1"))

tib_summed$DAST_dich <-tib_summed$DAST

tib_summed$DAST_dich[tib_summed$DAST_dich > 0] <- 1
summary(tib_summed$DAST_dich)

tib_summed$NODS_dich <- tib_summed$NODS
summary(tib_summed$NODS)
tib_summed$NODS_dich[tib_summed$NODS_dich > 0] <- 1
summary(tib_summed$NODS_dich)

#all variables that are needed for model specification
#VarBase$Demo, VarBase$Pnum
VarBase$Spec <- c('SSexp', 'SSdis', 'SSas', 'SSbor', 'BISa', 'BISm', 'BISn', 'Dhea', 'Dinv', 'Dgam', 'Deth', 'Drec', 'Dsoc', 'Dm', 'GBS', 'DAST_dich', 'FTND_dich', 'AUDIT', 'NODS_dich','SEC_1', 'SEC_2', 'SEC_3', 'SEC_4', 'SEC_5', 'SEC_6', 'SEC_7r', 'SEC_8', 'SEC_9', 'SEC_10', 'SEC_11', 'SEC_12')
VarBase$R <- c('SSexp', 'SSdis', 'SSas', 'SSbor', 'BISa', 'BISm', 'BISn', 'Dhea', 'Dinv', 'Dgam', 'Deth', 'Drec', 'Dsoc', 'Dm', 'GBS', 'DAST_dich', 'FTND_dich', 'AUDIT', 'NODS_dich')



# find DVs that are numeric / ordered factors, which can be converted / scaled
DVs <- NULL
for (i in 1:ncol(tib_summed)) {
  if (is.numeric(tib_summed[,i])) DVs <- c(DVs, names(tib_summed[i]))
  if (is.ordered(tib_summed[,i])) DVs <- c(DVs, names(tib_summed[i]))
}

tib_complete <- tib_summed[complete.cases(tib_summed),]


# create objects with standardized variables / residuals after controlling for age, sex,
tib_std <- subset(tib_complete, !Pnum%in%c(62, 254, 766, 617, 490, 671)) 
tib_std[] <- NA

#exclude the four participants that did not disclose their gender AND the two participants that reported "other" gender when computing residuals
residuals <- subset(tib_complete, !Pnum%in%c(62, 254, 766, 617, 490, 671)) 

key <- subset(tib_complete, !Pnum%in%c(62, 254, 766, 617, 490, 671)) 

residuals[] <- NA
residuals_num <- residuals
residuals_std <- residuals
residuals_std2 <- residuals

models <- list()
for (DV in DVs) {
  print(DV)
  curr_raw <- key[,DV]

 # convert measures to numeric
  if (is.numeric(curr_raw)) curr_raw_num <- curr_raw
  else curr_raw_num <- as.numeric(as.character(curr_raw))

  # standardize measures
  tib_std[,DV] <- scale(key[,DV])

 # get residuals (but not for "age")
# VarBase$Demo, VarBase$Pnum (key[,c(VarBase$Spec)])
  if (DV != "age" & DV != "Pnum" & DV != "sex") {
	formula <- as.formula(paste(DV, "~ 1 + age + sex"))


    if (is.numeric(curr_raw)) {
      p_col <- "grey"
      ind_ok <- which(!is.na(key[,DV]))
      mod.gauss <- lm(formula, data=key)
      models[[DV]] <- mod.gauss
      pred <- predict(mod.gauss, newdata=key[ind_ok,])
      residuals[ind_ok,DV] <- key[ind_ok,DV] - pred
      residuals_num[,DV] <- residuals[,DV]
      residuals_std[,DV] <- as.numeric(scale(residuals_num[,DV]))
      residuals_std2[,DV] <- as.numeric(scale(residuals_num[,DV]))
    } else {
      p_col <- "red"
      ind_ok <- which(!is.na(key[,DV]))
      mod.ord <- clm(formula, data=key)
      models[[DV]] <- mod.ord
      pred <- as.numeric(as.character(predict(mod.ord, type="class", newdata=key[ind_ok,])$fit))
      residuals[ind_ok,DV] <- key[ind_ok,DV] - pred
      residuals[,DV] <- as.ordered(residuals[,DV])
      residuals_num[,DV] <- as.numeric(as.character(residuals[,DV]))
      residuals_std[,DV] <- residuals[,DV]
      residuals_std2[,DV] <- as.numeric(scale(residuals_num[,DV]))
    }


  }
}

residuals[,c("age", "Pnum", "sex", "Pincome","Hincome")] <- key[,c("age", "Pnum", "sex", "Pincome","Hincome")]
residuals_num[,c("age", "Pnum", "sex", "Pincome","Hincome")] <- key[,c("age", "Pnum", "sex", "Pincome","Hincome")]
residuals_std[,c("age", "Pnum", "sex", "Pincome","Hincome")] <- key[,c("age", "Pnum", "sex", "Pincome","Hincome")]
residuals_std2[,c("age", "Pnum", "sex", "Pincome","Hincome")] <- key[,c("age", "Pnum", "sex", "Pincome","Hincome")]


save(tib_summed, tib_complete, key, tib_std, tib_filled, tib_cleaned, tib_imputed, VarBase, residuals,residuals_num,residuals_std,residuals_std2, file="C:/Users/dalla/Google Drive/R Coding/Comparisons_Risk/SEC_R_Crowd_2_22_2019_imputed.Rdata")

save(tib_complete, VarBase, residuals_std2, file="C:/Users/dalla/Google Drive/R Coding/Comparisons_Risk/SEC_R_Crowd_2_22_2019_analysis.Rdata")

save(dat_imp1, dat_imp2, dat_imp3, dat_imp4, dat_imp5, dat_imp6, dat_imp7, dat_imp8, dat_imp9, dat_imp10, dat_imp11, dat_imp12, dat_imp13, dat_imp14, dat_imp15, dat_imp16, dat_imp17, dat_imp18, dat_imp19, dat_imp20, tib_cleaned, tib_imputed, VarBase, file="C:/Users/dalla/Google Drive/R Coding/Comparisons_Risk/SEC_R_Crowd_2_22_2019_imputations.Rdata")

write.csv(tib_summed, file="C:/Users/dalla/Google Drive/R Coding/Comparisons_Risk/SEC_R_Crowd_2_22_2019_cleaned&imputed.csv")
